fix: split up call detail into multiple files
