+++ clients/web/src/components/calls/detail-pane/trace/popover/tool-tip/content.tsx
@@ -1,0 +1,11 @@
+import type { JSXElement } from "solid-js";
+import { Popover } from "@kobalte/core";
+
+export function Content(props: { children: JSXElement }) {
+  return (
+    <>
+      <Popover.Arrow />
+      <table>{props.children}</table>
+    </>
+  );
+}
